calculator/requirements.txt
Django-5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
whitenoise==6.7.0
